requirements.txt
Django==2.2.4
django-cors-headers==3.1.0
django-filter==2.2.0
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2019.2
sqlparse==0.3.0
